% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_reduced_matrix.R
\name{add_labels_to_reduced_matrix}
\alias{add_labels_to_reduced_matrix}
\title{Run all steps to create a filled worker matrix
#'}
\usage{
add_labels_to_reduced_matrix(reduced_matrix, old_matrix, sic = "sic07")
}
\arguments{
\item{reduced_matrix}{A matrix}

\item{industry_data}{A data frame wit the indusry classification.}

\item{ind_class}{The label for the industry classification}
}
\value{
An updated matrix.
}
\description{
Run all steps to create a filled worker matrix
#'
}
\examples{
\dontrun{
fill_industry_transition_matrix_from_reduced <- function(sample, cuts, cutsName,indClass)
}
}
