% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ashe_sic_5dig}
\alias{ashe_sic_5dig}
\title{Synthetic ASHE dataset}
\format{A data frame with 1301 rows and 133 variables.}
\source{
Synthetic generated
}
\usage{
ashe_sic_5dig
}
\description{
Sample generated from expected columns on the ASHE dataset, contains all missing values except the
sic03 and sic07 variables containing the industrial classification.
}
\keyword{datasets}
