% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashe_bres_ratio.R
\name{check_ashe_bres_ratio}
\alias{check_ashe_bres_ratio}
\title{Fill wotker transition matrix from a vector of worker transitions}
\usage{
check_ashe_bres_ratio(ashe, bres, year)
}
\arguments{
\item{workermatrix}{A data frame.}

\item{transitionmatrix}{The personal id of the worker.}
}
\value{
An updated matrix.
}
\description{
TODO: For a particular worker fill the transition matrix.
}
\examples{
\dontrun{
fill_worker_transition_matrix(transitionmatrix,workermatrix)
}
}
