% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_and_fill_industry_transition_matrix.R
\name{create_and_fill_industry_transition_matrix}
\alias{create_and_fill_industry_transition_matrix}
\title{Run all steps to create a filled worker matrix}
\usage{
create_and_fill_industry_transition_matrix(sample, cuts, cutsString,
  indClass)
}
\arguments{
\item{sample}{A data frame.}

\item{cuts}{The cuts to be applied to the sample}
}
\value{
An updated matrix.
}
\description{
TODO: For a particular sample run cuts and fill the transition matrix.
}
\examples{
\dontrun{
create_and_fill_industry_transition_matrix <- function(sample, cuts, cutsString,indClass)
}
}
