% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_normalised_transition_matrix.R
\name{create_normalised_transition_matrix}
\alias{create_normalised_transition_matrix}
\title{Run all steps to create a filled worker matrix
#'}
\usage{
create_normalised_transition_matrix(transition_matrix)
}
\arguments{
\item{transition_matrix}{A matrix}
}
\value{
A matrix.
}
\description{
Run all steps to create a filled worker matrix
#'
}
\examples{
\dontrun{
create_normalised_transition_matrix <- function(transition_matrix)
}
}
