% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_normalised_transition_matrix_from_reduced.R
\name{create_normalised_transition_matrix_from_reduced}
\alias{create_normalised_transition_matrix_from_reduced}
\title{Run all steps to create a normalised worker matrix from an input raw count matrix
#'}
\usage{
create_normalised_transition_matrix_from_reduced(reduced_matrix, mincuts,
  algorithm, sic = "sic07")
}
\arguments{
\item{transition_matrix}{A matrix}
}
\value{
A matrix.
}
\description{
Run all steps to create a normalised worker matrix from an input raw count matrix
#'
}
\examples{
\dontrun{
create_normalised_transition_matrix_from_reduced <- function(reduced_matrix,mincuts,algorithm,sic='sic07')
}
}
