% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_null_model_transition_matrix.R
\name{create_null_model_transition_matrix}
\alias{create_null_model_transition_matrix}
\title{Starting from input adjacency matrix, create a null hypothesis matrix based on Nefkke method.
#'}
\usage{
create_null_model_transition_matrix(transition_matrix)
}
\arguments{
\item{transition_matrix}{An adjacency matrix}
}
\value{
A matrix.
}
\description{
Starting from input adjacency matrix, create a null hypothesis matrix based on Nefkke method.
#'
}
\examples{

create_null_model_transition_matrix(transition_matrix)

}
