% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reduced_transition_matrix.R
\name{create_reduced_transition_matrix}
\alias{create_reduced_transition_matrix}
\title{Convert worker transition matrix to a 3-column form}
\usage{
create_reduced_transition_matrix(transitionmatrix, min_counts)
}
\arguments{
\item{transitionmatrix}{A matrix.}

\item{min_counts}{Minimum number of counts in the matrix to be considered}
}
\value{
An 3 column x N rows matrix (N is number of flows)
}
\description{
Convert worker transition matrix to a 3-column form
}
\examples{
\dontrun{
create_reduced_transition_matrix(transitionmatrix, min_counts)
}
}
