% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_industry_transition_counts_matrix.R
\name{empty_industry_transition_counts_matrix}
\alias{empty_industry_transition_counts_matrix}
\title{Create an empty industry transition counts matrix}
\usage{
empty_industry_transition_counts_matrix(data)
}
\arguments{
\item{data}{A data frame.}

\item{colname}{The name of the industry code column in the data.}
}
\value{
An empty matrix.
}
\description{
TODO: Produce a empty matrix for all industry classification
}
\examples{
\dontrun{
empty_industry_transition_counts_matrix(ashe_sic_5dig, colname = "sic07")
}
}
