% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_transition_matrix.R
\name{empty_transition_matrix}
\alias{empty_transition_matrix}
\title{Create an empty worker transition matrix}
\usage{
empty_transition_matrix(data, colname)
}
\arguments{
\item{data}{A data frame.}

\item{colname}{The name of the industry code column in the data.}
}
\value{
An empty matrix.
}
\description{
TODO: description
}
\examples{
\dontrun{
empty_transition_matrix(ashe_sic_5dig, colname = "sic07")
}
}
