% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_worker_transition_matrix.R
\name{fill_worker_transition_matrix}
\alias{fill_worker_transition_matrix}
\title{Fill wotker transition matrix from a vector of worker transitions}
\usage{
fill_worker_transition_matrix(transitionmatrix, workermatrix)
}
\arguments{
\item{transitionmatrix}{The personal id of the worker.}

\item{workermatrix}{A data frame.}
}
\value{
An updated matrix.
}
\description{
TODO: For a particular worker fill the transition matrix.
}
\examples{
\dontrun{
fill_worker_transition_matrix(ashe_sic_5dig, piden = 1)
}
}
