% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all steps to create a reduced matrix}
\usage{
run_all(input_path, output_path, log_path, colname, cuts, labelcut,
  min_counts)
}
\arguments{
\item{input_path}{Input path.}

\item{output_path}{Input path.}

\item{log_path}{Path to write output}

\item{colname}{The name of the industry code column in the data.}

\item{cuts}{The cuts to be applied to the sample}

\item{min_counts}{Labels to name the outputs}

\item{labelcuts}{Labels to name the outputs}
}
\value{
A data frame
}
\description{
TODO: For a particular worker id, produce a vector of all transitions.
}
\examples{
\dontrun{
run_all(ashe_sic_5dig, piden = 1)
}
}
