% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ashe_sample_summary.R
\name{run_ashe_sample_summary}
\alias{run_ashe_sample_summary}
\title{Sample summary values}
\usage{
run_ashe_sample_summary(sample, name = "", selection_cuts = "")
}
\arguments{
\item{sample}{A data frame.}
}
\value{
A dictionary
}
\description{
TODO: For a particular sample, get all general diagnostics.
}
\examples{
\dontrun{
run_ashe_sample_summary(sample)
}
}
