% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_cuts_ashe.R
\name{selection_custs_ashe}
\alias{selection_custs_ashe}
\title{Selection cut function for ASHE dataset}
\usage{
selection_custs_ashe(data, cut)
}
\arguments{
\item{data}{A data frame.}

\item{cut}{string
The selection cut requested.}
}
\value{
A data frame.
}
\description{
TODO: Produce a empty matrix for all industry classification
}
\examples{
\dontrun{
selection_custs_ashe(ashe_sic_5dig, piden>1 && sex=="F")
}
}
