% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sic03_07_5dig}
\alias{sic03_07_5dig}
\title{SIC code dataset for 2003 and 2007 (5 digit)}
\format{A data frame with 1301 rows and 5 variables.}
\source{
\url{http://www.ons.gov.uk/ons/guide-method/classifications/current-standard-classifications/standard-industrial-classification/correlation-between-uk-sic-2003-and-2007.xls}
}
\usage{
sic03_07_5dig
}
\description{
Current Standard Industrial Classification (SIC) used in classifying business
establishments and other statistical units by the type of economic activity in which they are engaged.
}
\keyword{datasets}
