% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_transition_matrix.R
\name{worker_transition_matrix}
\alias{worker_transition_matrix}
\title{Create a vector of worker transitions}
\usage{
worker_transition_matrix(data, piden, colname)
}
\arguments{
\item{data}{A data frame.}

\item{piden}{The personal id of the worker.}

\item{colname}{The name of the industry code column in the data.}
}
\value{
An list of pairs.
}
\description{
TODO: For a particular worker id, produce a vector of all transitions.
}
\examples{
\dontrun{
worker_transition_matrix(ashe_sic_5dig, piden = 1)
}
}
