% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_labels_to_reduced_matrix.R
\name{add_labels_to_reduced_matrix}
\alias{add_labels_to_reduced_matrix}
\title{Augment a reduced matrix with label information and transition counts from original dataset}
\usage{
add_labels_to_reduced_matrix(reduced_matrix, previous_matrix,
  ind_class = "sic07")
}
\arguments{
\item{reduced_matrix}{A matrix with 3 columns from industry transitions}

\item{previous_matrix}{A data frame with the raw counts transition daya}

\item{ind_class}{The label for the industry classification}
}
\value{
An updated matrix.
}
\description{
Augment a reduced matrix with label information and transition counts from original dataset
}
\examples{

add_labels_to_reduced_matrix(reduced_matrix,old_matrix,'sic07')

}
