% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_all_normalised_transition_matrix.R
\name{create_all_normalised_transition_matrix}
\alias{create_all_normalised_transition_matrix}
\title{Run all steps to create all the normalised worker matrices from the secure lab inputs
#'}
\usage{
create_all_normalised_transition_matrix(input_path = "")
}
\arguments{
\item{the}{input path where all the secure lab outputs are
No input}
}
\value{
No output all the matrices get writen in the python/labour_flows_matrix/data/ directory
}
\description{
Run all steps to create all the normalised worker matrices from the secure lab inputs
#'
}
\examples{
\dontrun{
create_all_normalised_transition_matrix <- function(input_path)
}
}
