% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/create_normalised_transition_matrix_from_reduced.R
\name{create_normalised_transition_matrix_from_reduced}
\alias{create_normalised_transition_matrix_from_reduced}
\title{Run all steps to create a filled worker matrix
#'}
\usage{
create_normalised_transition_matrix_from_reduced(reduced_matrix,
  labelmatrix, mincuts)
}
\arguments{
\item{transition_matrix}{A matrix}
}
\value{
A matrix.
}
\description{
Run all steps to create a filled worker matrix
#'
}
\examples{
\dontrun{
create_normalised_transition_matrix <- function(transition_matrix)
}
}
