% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_industry_transition_matrix_from_reduced.R
\name{fill_industry_transition_matrix_from_reduced}
\alias{fill_industry_transition_matrix_from_reduced}
\title{Read a reduced transition matrix and build an ND x ND adjacency matrix
#'}
\usage{
fill_industry_transition_matrix_from_reduced(reduced_matrix, algorithm)
}
\arguments{
\item{reduced_matrix}{A matrix}

\item{algorithm}{The algorithm used to build the matrix (in order to decide symetrisation strategy)}
}
\value{
An updated matrix.
}
\description{
Read a reduced transition matrix and build an ND x ND adjacency matrix
#'
}
\examples{
\dontrun{
fill_industry_transition_matrix_from_reduced <- function(reduced_matrix,algorithm)
}
}
