% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_industry_transition_matrix_from_reduced.R
\name{fill_industry_transition_matrix_from_reduced}
\alias{fill_industry_transition_matrix_from_reduced}
\title{Run all steps to create a filled worker matrix
#'}
\usage{
fill_industry_transition_matrix_from_reduced(reduced_matrix, industry_data,
  ind_class)
}
\arguments{
\item{reduced_matrix}{A matrix}

\item{industry_data}{A data frame wit the indusry classification.}

\item{ind_class}{The label for the industry classification}
}
\value{
An updated matrix.
}
\description{
Run all steps to create a filled worker matrix
#'
}
\examples{
\dontrun{
fill_industry_transition_matrix_from_reduced <- function(sample, cuts, cutsName,indClass)
}
}
