% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_all.R
\name{run_all}
\alias{run_all}
\title{Run all steps to create a reduced matrix}
\usage{
run_all(selected_sample, output_path, colname, min_counts)
}
\arguments{
\item{output_path}{Input path.}

\item{colname}{The name of the industry code column in the data.}

\item{min_counts}{Labels to name the outputs}

\item{input_path}{Input path.}

\item{cuts}{The cuts to be applied to the sample}

\item{labelcuts}{Labels to name the outputs}

\item{log_path}{Path to write output}
}
\value{
Nothing
}
\description{
Run all steps to create a reduced matrix
}
\examples{
\dontrun{
run_all(input_path, output_path, log_path, colname, cuts, labelcut, min_counts)
}
}
